Initial separation of rendering and game logic
Untested currently, but this is the first step.
