+++ src/gameComponents/Squadrons/Squadron.java
@@ -22,9 +22,6 @@ import java.util.ArrayList;
  */
 public class Squadron implements GameComponent {
 
-    private static ShaderProgram WHITE_SOLID = ShaderPrograms.buildSolidColorShader("white");
-    private static ShaderProgram BLACK_SOLID = ShaderPrograms.buildSolidColorShader("black");
-
     private final String type;
     private final String name;
     private final boolean unique;
@@ -37,23 +34,9 @@ public class Squadron implements GameComponent {
     private final ArrayList<String> keywords;
     private final int pointsValue;
     private ArrayList<DefenseToken> defenseTokens;
-    private GameItemSorter gameItems = new GameItemSorter();
     private BBDPoint currentLocation = new BBDPoint(0,0);
-    private boolean renderSquadrons;
 
 
-    /**
-     * Static objects to build meshes.  Long term these will be moved somewhere else once rendering is separated from the Squadron object.
-     */
-    private static BBDPolygon plasticBase = GeometryGenerators.createNGon(new BBDPoint(0,0), GameConstants.SQUADRON_PLASTIC_RADIUS, 100);
-    private static float[] plasticPositions = Mesh.buildMeshPositions(plasticBase);
-    private static float[] plasticTex = Mesh.buildTextureCoordinates(plasticBase);
-    private static int[] plasticIndices = Mesh.buildIndices(plasticBase);
-
-    private static BBDPolygon cardboard = GeometryGenerators.createNGon(new BBDPoint(0,0), GameConstants.SQUADRON_CARDBOARD_RADIUS, 100);
-    private static float[] cardboardPositions = Mesh.buildMeshPositions(cardboard);
-    private static float[] cardboardTex = Mesh.buildTextureCoordinates(cardboard);
-    private static int[] cardboardIndices = Mesh.buildIndices(cardboard);
 
 
     /**
@@ -106,11 +89,6 @@ public class Squadron implements GameComponent {
         this.keywords = original.keywords;
         this.pointsValue = original.pointsValue;
         this.defenseTokens = original.defenseTokens;
-        this.renderSquadrons = renderSquadrons;
-
-        if (renderSquadrons) {
-            this.buildGameItems();
-        }
     }
 
     @Override
@@ -139,38 +117,6 @@ public class Squadron implements GameComponent {
     }
 
     /**
-     * Build the GameItem objects to be used to render this object.
-     */
-    private void buildGameItems(){
-        GameItem2d plasticBaseItem = new GameItem2d(new Mesh(plasticPositions, plasticTex, plasticIndices, null), WHITE_SOLID, plasticBase, GameConstants.LAYER_SQUADRON_PLASTIC, false);
-        GameItem2d cardboardItem = new GameItem2d(new Mesh(cardboardPositions, cardboardTex, cardboardIndices, null), BLACK_SOLID, cardboard, GameConstants.LAYER_SQUADRON_CARDBOARD, false);
-
-        BBDPolygon poly = GeometryGenerators.buildQuad(20, 20);
-        ShaderProgram shader = ShaderPrograms.TEXTURED_GENERIC;
-        Texture texture = new Texture("assets/images/squadrons/squad_"+buildSquadFileName());
-        GameItem2d squadronGraphic = new GameItem2d(Mesh.buildMeshFromPolygon(poly, texture), shader, poly, GameConstants.LAYER_SQUADRON_GRAPHIC, false);
-        this.gameItems.addItems(squadronGraphic);
-        this.gameItems.addItems(cardboardItem);
-        this.gameItems.addItems(plasticBaseItem);
-    }
-
-    /**
-     * Build a string based on the squadron object's properties to grab the appropriate descriptively named file.
-     * Concatenates a few fields and cleans up outstanding chars like spaces, quotes etc.
-     * @return image file to be used from the assets directory
-     */
-    public String buildSquadFileName() {
-        String baseFileName = this.type;
-        if(this.unique){
-            baseFileName = baseFileName+"_"+this.name;
-        }
-
-        String cleanedFileName = baseFileName.toLowerCase().replace(" ", "-").replace("\"", "");
-
-        return cleanedFileName+".png";
-    }
-
-    /**
      * Root movement function.  All movement functions eventually lead to here.  Function is private because it is the one
      * that modifies items of the class.
      * @param newPoint new BBDPoint to use for the location of the squadron
@@ -179,9 +125,6 @@ public class Squadron implements GameComponent {
         this.currentLocation = newPoint;
         float newX = newPoint.getXLoc();
         float newY = newPoint.getYLoc();
-        for(int i =0; i<gameItems.getItemCount(); i++){
-            gameItems.getItem(i).setPosition(newX, newY, gameItems.getItem(i).getPosition().z);
-        }
     }
 
     /**
@@ -258,11 +201,23 @@ public class Squadron implements GameComponent {
         return defenseTokens;
     }
 
-    public GameItemSorter getGameItems() {
-        return gameItems;
-    }
-
     public BBDPoint getLocation(){
         return this.currentLocation;
+    }
+
+    /**
+     * Build a string based on the squadron object's properties to grab the appropriate descriptively named file.
+     * Concatenates a few fields and cleans up outstanding chars like spaces, quotes etc.
+     * @return image file to be used from the assets directory
+     */
+    public String buildSquadFileName() {
+        String baseFileName = this.type;
+        if(this.unique){
+            baseFileName = baseFileName+"_"+this.name;
+        }
+
+        String cleanedFileName = baseFileName.toLowerCase().replace(" ", "-").replace("\"", "");
+
+        return cleanedFileName+".png";
     }
 }
